Initialization of repository
TODO: Fix the issue with the decision tree - one situation where the machines makes a move allowing the opponent to make ever winning move.

Probably add additional parameter to hold the result of the game in the leaf nodes.
Also, the decision trees might be predefined and loaded in when the game starts, based on initial conditions. Then, as the game continues, the decision tree residing in the memory would become a subtree of the original one.